Add files via upload
Modifications made to file organization.  Separate file for each class (currently only one class, for the data to be transmitted to the web server) and one project file with all the code to retrieve the sensor values from the Raspberry Pi and assign those values to the fields of the data record.